+++ src/magnetic_field.c
@@ -32,7 +32,7 @@ void Make_magnetic_field()
 
 static void set_magnetic_vector_potential()
 {
-	const float boxhalf = 0.5 * Param.Boxsize;
+	const double boxhalf = 0.5 * Param.Boxsize;
 
 		#pragma omp parallel for 
     	for (size_t ipart = 0; ipart < Param.Npart[0]; ipart++) {
@@ -44,7 +44,7 @@ static void set_magnetic_vector_potential()
 			if (Halo[i].Mass[0] == 0) // DM only halos
 				continue;
 
-			float dx = P[ipart].Pos[0] - Halo[i].D_CoM[0] - boxhalf,
+			double dx = P[ipart].Pos[0] - Halo[i].D_CoM[0] - boxhalf,
 				  dy = P[ipart].Pos[1] - Halo[i].D_CoM[1] - boxhalf,
 				  dz = P[ipart].Pos[2] - Halo[i].D_CoM[2] - boxhalf;
 
@@ -60,9 +60,9 @@ static void set_magnetic_vector_potential()
 				A_max = A;
 			}
 		
-			SphP[ipart].Apot[0] = (float) A_max;
-			SphP[ipart].Apot[1] = (float) A_max;
-	    	SphP[ipart].Apot[2] = (float) A_max;
+			SphP[ipart].Apot[0] = (double) A_max;
+			SphP[ipart].Apot[1] = (double) A_max;
+	    	SphP[ipart].Apot[2] = (double) A_max;
 		}
 
     return ;  
@@ -70,7 +70,7 @@ static void set_magnetic_vector_potential()
 
 static void normalise_magnetic_field()
 {
-	const float boxhalf = 0.5 * Param.Boxsize;
+	const double boxhalf = 0.5 * Param.Boxsize;
 
  	double max_B2 = 0;
 
@@ -101,7 +101,7 @@ static void normalise_magnetic_field()
 		double B2 = p2(SphP[ipart].Bfld[0]) + p2(SphP[ipart].Bfld[1]) + 
 					p2(SphP[ipart].Bfld[2]);
 	
-		float x = P[ipart].Pos[0] - boxhalf,
+		double x = P[ipart].Pos[0] - boxhalf,
 			  y = P[ipart].Pos[1] - boxhalf,
 			  z = P[ipart].Pos[2] - boxhalf;
